fix(minio): remove default bucket lifecycle rule (#30)
Because

- Some distributions might want to have objects with no expiration.
E.g.: Instill Cloud on the Enterprise subscription.
- The 30 day bucket expiration rule was implemented before tag-based
lifecycle management, so it isn't required anymore. This will be applied
to tags that have no associated lifecycle rule.

This commit

- Removes the default 30 day expiration rule.
- The new configuration without the default expiration will override the
existing one. Therefore, no migration is required to remove the tag and
its associated rule. Clients that update this package will change the
MinIO bucket settings on startup.